Pulling index block retrying out to `Messages` class.
