+++ graylog2-server/src/test/java/org/graylog2/indexer/messages/MessagesRetryWaitTest.java
@@ -14,11 +14,9 @@
  * You should have received a copy of the GNU General Public License
  * along with Graylog.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.graylog.storage.elasticsearch6;
+package org.graylog2.indexer.messages;
 
 import com.github.rholder.retry.WaitStrategy;
-import org.graylog2.indexer.messages.IndexBlockRetryAttempt;
-import org.graylog2.indexer.messages.Messages;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -27,7 +25,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 public class MessagesRetryWaitTest {
     @Test
     void secondsBasedRetryWaitsForSecondsStartingWith1() {
-        WaitStrategy waitStrategy = MessagesAdapterES6.exponentialWaitSeconds;
+        WaitStrategy waitStrategy = Messages.exponentialWaitSeconds;
         assertAll(
                 () -> assertThat(waitStrategy.computeSleepTime(new IndexBlockRetryAttempt(1))).isEqualTo(1000),
                 () -> assertThat(waitStrategy.computeSleepTime(new IndexBlockRetryAttempt(2))).isEqualTo(2000),
@@ -40,7 +38,7 @@ public class MessagesRetryWaitTest {
     // This test was added to document how the retry strategy actually behaves since this is hard to deduct from the code
     @Test
     void millisecondsBasedRetryWaitsForMillisecondsStartingWith2() {
-        WaitStrategy waitStrategy = MessagesAdapterES6.exponentialWaitMilliseconds;
+        WaitStrategy waitStrategy = Messages.exponentialWaitMilliseconds;
         assertAll(
                 () -> assertThat(waitStrategy.computeSleepTime(new IndexBlockRetryAttempt(1))).isEqualTo(2),
                 () -> assertThat(waitStrategy.computeSleepTime(new IndexBlockRetryAttempt(2))).isEqualTo(4),
